Fix wrong pc type

+++ tests/unit/test_x86.c
@@ -1619,7 +1619,7 @@ static void test_x86_vtlb(void)
     uc_hook hook;
     char code[] = "\xeb\x02\x90\x90\x90\x90\x90\x90"; // jmp 4; nop; nop; nop;
                                                       // nop; nop; nop
-    uint64_t r_eip = 0;
+    uint32_t r_eip = 0;
 
     uc_common_setup(&uc, UC_ARCH_X86, UC_MODE_32, code, sizeof(code) - 1);
